#8228
fixed the odbc connection module build on old gcc (4.2.1 on freebsd 9.3 and
freebsd 7.4 i386), where charstring::parseNumber(SQLUINTEGER) was ambiguous at
odbc.cpp:1895, 1909, 1977, 1990 (all in odbcconnection::getDatabaseFeatures()).
parseNumber has uint32_t and uint64_t overloads; SQLUINTEGER on those platforms
matches neither exactly, so the old compiler could not pick one.  cast the
argument to (uint32_t) - SQLUINTEGER is a 32-bit unsigned, so this is the
exact-width, lossless match - which makes overload resolution unambiguous.
verified on freebsd9x86 (gcc 4.2.1, unixODBC 2.3.1): "gmake odbc.lo" now
compiles clean (previously "call of overloaded parseNumber(SQLUINTEGER&) is
ambiguous").

+++ src/connections/odbc.cpp
@@ -1892,7 +1892,7 @@ const char * const *odbcconnection::getDatabaseFeatures() {
 	SQLGetInfo(dbc,SQL_MAX_BINARY_LITERAL_LEN,
 					&uintbuf,sizeof(uintbuf),&size);
 	databasefeatures[FEATURE_MAX_BINARY_LITERAL_LENGTH]=
-		charstring::parseNumber(uintbuf);
+		charstring::parseNumber((uint32_t)uintbuf);
 
 	// SQL_MAX_CATALOG_NAME_LEN -> usmallint
 	usmallintbuf=0;
@@ -1906,7 +1906,7 @@ const char * const *odbcconnection::getDatabaseFeatures() {
 	SQLGetInfo(dbc,SQL_MAX_CHAR_LITERAL_LEN,
 					&uintbuf,sizeof(uintbuf),&size);
 	databasefeatures[FEATURE_MAX_CHAR_LITERAL_LENGTH]=
-		charstring::parseNumber(uintbuf);
+		charstring::parseNumber((uint32_t)uintbuf);
 
 	// SQL_MAX_COLUMN_NAME_LEN -> usmallint
 	usmallintbuf=0;
@@ -1974,7 +1974,7 @@ const char * const *odbcconnection::getDatabaseFeatures() {
 	uintbuf=0;
 	SQLGetInfo(dbc,SQL_MAX_INDEX_SIZE,&uintbuf,sizeof(uintbuf),&size);
 	databasefeatures[FEATURE_MAX_INDEX_LENGTH]=
-		charstring::parseNumber(uintbuf);
+		charstring::parseNumber((uint32_t)uintbuf);
 
 	// SQL_MAX_PROCEDURE_NAME_LEN -> usmallint
 	usmallintbuf=0;
@@ -1987,7 +1987,7 @@ const char * const *odbcconnection::getDatabaseFeatures() {
 	uintbuf=0;
 	SQLGetInfo(dbc,SQL_MAX_ROW_SIZE,&uintbuf,sizeof(uintbuf),&size);
 	databasefeatures[FEATURE_MAX_ROW_SIZE]=
-		charstring::parseNumber(uintbuf);
+		charstring::parseNumber((uint32_t)uintbuf);
 
 	// SQL_MAX_OWNER_NAME_LEN -> usmallint
 	usmallintbuf=0;
